passing tests (to-do completed)

+++ src/DataTypes.java
@@ -1,11 +1,9 @@
+import java.util.ArrayList;
 import java.util.List;
 
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
-
-        int s = 0;
-        // below is a "foreach" loop which iterates through numbers
+        long s = 0;
         for (int x : numbers) {
             s += x;
         }
